Adjust assignment of Bash as SHELL for FreeBSD
Some testing I've been doing involving DFXML occurs in FreeBSD
environments, where Bash is not installed in the base operating system,
and is installed to `/usr/local/bin/bash`.  This is why the Makefiles
have used `?=` assignment here.  Unfortunately, in Github Actions and in
FreeBSD, `?=` sees SHELL in the calling Make environment and does not
act on the backup value.  (I had not noticed this for a while because I
had other remediations, such as soft-linking `/bin/bash` or relying on
my active shell being Bash.  So, we have another case of CI revealing
operation assumptions)

Tests to date have not shown any need for Bash over Sh*, but tests I'm
proposing in the next patch will rely on Bash's `source` being available
to use virtual environments.

*(One potential case where we would have run into this was related to
Issue 39, but that failed before Bash detection for a different reason
that is also remediated in the next patch.)

I'm not sure we'll be able to test this in a FreeBSD environment with
Github Actions---from runner Issue 385, it looks like FreeBSD isn't
supported---but I can at least confirm I've replicated the behaviors I
described above, and confirmed this change works, in a FreeBSD system I
have handy.

References:
* [DFXML Issue 39] https://github.com/simsong/dfxml/issues/39
* [runner Issue 385] https://github.com/actions/runner/issues/385

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>