Move prior venv test to new top-level testing strategy
This patch implements some practices I've found helpful for developing
command-line tools for the CASE community.  See e.g. the tooling
available in CASE-Utilities-Python, the code base that relates to the
CASE ontology close to the same way this code base relates to the DFXML
schema.

There was a test reported as causing a problem in Issue 39, related to
building virtual environments.  Since writing that test, I've come
across this experience:
1. The Python module venv is different from the module virtualenv.
   venv is built in to the Python environment installed in Windows and
   macOS.  (Ubuntu requires an extra package be installed for either
   venv or virtualenv.)  venv meets the testing need for a module to
   install with `pip`.
   1.1. I've found installing the module as part of CI is an
        elementarily necessary test of functionality.  For instance,
        the directory rename that was just merged missed applying part
        of the rename in `/setup.py`, and this broke the module build.
        I'll fix that in the following patch, but wanted to trigger a CI
        failure to demonstrate.
2. I've learned a bit more about `pytest` and fixtures, so I'd be more
   comfortable migrating some of the prior tests to `pytest`.  I've done
   so in this patch with what was the "simple read" test.

I'd like tests to start being separated from source trees, and moved
into a directory under `/tests` that can have a parallel structure to
the top source directory.  I believe this will make some of the testing
related to this repository's new command-line tools easier to organize,
but this is a question of file organization style, a matter of taste.

I think we will be able to more cleanly delinate tests that are expected
to exhibit and exercise command-line behaviors from tests of internal
module functionality.  I'd like command-line tests to run from `make`,
and module-behavior tests to run from `pytest`.  I know Simson has
expressed a desire in the past for tests to run from top-directory call
to `pytest` - I'm proposing a compromise with this patch, that `pytest`
be called from inside the `/tests` directory so we can guarantee first
the `dfxml` module is installed and available for test scripts.

References:
* [CASE-Utilities-Python] https://github.com/casework/CASE-Utilities-Python
* [Issue 39] https://github.com/simsong/dfxml/issues/39

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>