Use setup.cfg
setup.cfg will enable an alternative mechanism for storing the version
of the overall DFXML package, to be shown in a following patch.

This initial patch confirms transferring functionality from setup.py to
setup.cfg does not cause any CI failures.

FEATURE TEMPORARY REMOVAL:
This patch removes `idifference.py` and `rdifference.py` from
installation into the environment's `.../bin` directory.  This does not
cause a failure of CI, but may be an unexpected regression.  This is a
confluence of Pull Request 8 and setup.cfg expecting console entry
points to be under the package directory.

I recall that there is a way to tell setup.cfg to also refer to a
directory aside from the top package directory, but I can't remember
where that documentation is.

References:
* [Pull Request 8] https://github.com/dfxml-working-group/dfxml_python/pull/8

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>