Run __eq__ type checks with consistent pattern
This implements suggestions from mypy while preserving a custom behavior
of the Objects module.

The custom behavior I'd like to preserve is a `TypeError` being thrown
when an `Objects.XObject` is compared against an object outside of the
class hierarchy.

mypy requires a pattern for the type signature of `XObject.__eq__` that
conforms to the Liskov substitution principal.  So, `other` must have
type `object`.  mypy can also recognize `isinstance()`, so it suggests
adding a `if not isinstance(other, TypeYouWant): return False`.  But,
just using mypy's suggestion catches `None`, which is the one type and
value outside the hierarchy where it will be generally be fine to report
inequality without raising a `TypeError`.

The result of the above reasoning is three-step checking in each
`__eq__` method.
1. None? Ok, return False.
2. Unexpected type? Raise a runtime error like in the setter methods.
3. Run an isinstance() check that, while it feels redudant, appeases
   mypy's static type-checking.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>