Reorder class definitions - ByteRuns up front
The following patch will be annotating the various `byte_runs` property
getters with the ByteRuns type.  If the ByteRuns class is not defined
sooner, either mypy or unit tests fail on encountering ByteRuns,
considering it an undefined symbol.

This may be a portent of further class reordering to occur as more type
annotations are added.

Note that though the `git diff` output is pretty chaotic, all that
occurred in this patch is moving `ByteRun` and `ByteRuns` to be before
`DiskImageObject`, with no other changes occurring.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>