Fix NameError thrown in ignorable_name()
Fix NameError-exception in the ignorable_name()-function in
idifference2.py, which stemmed from introducing a separate
main()-function with the intent to install the script via
setup.cfg's console_scripts in the future. By doing so
`args` was not available in the module-scope anymore, which
led to the a/m exception. This is corrected by setting a
module-wide variable.