Remove order-randomness in property serializations
The Class-level `_class_properties` sequences were initially defined as
sets.  This causes some randomness when iterating over the set, as their
member order is not guaranteed.  The result is non-deterministic output
that parses fine but causes unnecessary Git noise if results are being
pre-computed and stored.

This patch revises the `_class_properties` sequences to be Lists instead
of Sets.  According to the unit testing, some "Jitter" that was
previously happening in some serialization is no longer happening, and
the only impacted code outside of objects.py was in the `walk_to_dfxml`
script.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>