Document `pre-commit` hook
Text adapted from contributions to `case-utils` by @kchason.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* https://github.com/casework/CASE-Utilities-Python/pull/37

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>