Apply Black
This patch applies effects of the following patch that adds `black` to
CI.

The patch-ordering is because adding to CI first would cause CI to fail
on finding files need to be re-formatted by `pre-commit`.  This also
separates manual changes from automated changes.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>