Enable isort compatibility with Black and apply changes
The one manual change in this patch is to `/.pre-commit-config.yaml`.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.